Add blink got
